pyproject.toml
[project]
name = "docrestest"
version = "0.1.0"
description = ""
authors = [
    {name = "Zerrant",email = "Kao242003@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "accelerate (==1.9.0)",
    "annotated-types (==0.7.0)",
    "anyio (==4.10.0)",
    "argon2-cffi (==25.1.0)",
    "argon2-cffi-bindings (==25.1.0)",
    "arrow (==1.3.0)",
    "asttokens (==3.0.0)",
    "async-lru (==2.0.5)",
    "attrs (==25.3.0)",
    "babel (==2.17.0)",
    "backports-tarfile (==1.2.0)",
    "beautifulsoup4 (==4.13.4)",
    "bleach (==6.2.0)",
    "build (==1.3.0)",
    "cachecontrol (==0.14.3)",
    "certifi (==2025.8.3)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.2)",
    "cleo (==2.1.0)",
    "colorama (==0.4.6)",
    "contourpy (==1.3.2)",
    "crashtest (==0.4.1)",
    "cryptography (==45.0.6)",
    "cycler (==0.12.1)",
    "debugpy (==1.8.15)",
    "decorator (==5.2.1)",
    "decord (==0.6.0)",
    "defusedxml (==0.7.1)",
    "distlib (==0.4.0)",
    "distro (==1.9.0)",
    "dulwich (==0.22.8)",
    "einops (==0.8.1)",
    "exceptiongroup (==1.3.0)",
    "executing (==2.2.0)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.13.1)",
    "findpython (==0.6.3)",
    "fonttools (==4.59.0)",
    "fqdn (==1.5.1)",
    "fsspec (==2024.6.1)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "huggingface-hub (==0.34.3)",
    "idna (==3.10)",
    "imageio (==2.37.0)",
    "importlib-metadata (==8.7.0)",
    "installer (==0.7.0)",
    "ipython (==8.37.0)",
    "isoduration (==20.11.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.3.0)",
    "jedi (==0.19.2)",
    "jinja2 (==3.1.4)",
    "jiter (==0.10.0)",
    "json5 (==0.12.0)",
    "jsonpointer (==3.0.0)",
    "jsonschema (==4.25.0)",
    "jsonschema-specifications (==2025.4.1)",
    "jupyter-events (==0.12.0)",
    "jupyter-lsp (==2.2.6)",
    "jupyter-client (==8.6.3)",
    "jupyter-core (==5.8.1)",
    "jupyter-server (==2.17.0)",
    "jupyter-server-terminals (==0.5.3)",
    "jupyterlab-pygments (==0.3.0)",
    "keyring (==25.6.0)",
    "kiwisolver (==1.4.8)",
    "lark (==1.2.2)",
    "lazy-loader (==0.4)",
    "markupsafe (==2.1.5)",
    "matplotlib (==3.10.5)",
    "matplotlib-inline (==0.1.7)",
    "mistune (==3.1.3)",
    "more-itertools (==10.7.0)",
    "mpmath (==1.3.0)",
    "msgpack (==1.1.1)",
    "nbclient (==0.10.2)",
    "nbconvert (==7.16.6)",
    "nbformat (==5.10.4)",
    "nest-asyncio (==1.6.0)",
    "networkx (==3.4.2)",
    "numpy (==2.2.6)",
    "openai (==1.100.2)",
    "opencv-python-headless (==4.11.0.86)",
    "overrides (==7.7.0)",
    "packaging (==25.0)",
    "pandocfilters (==1.5.1)",
    "parso (==0.8.4)",
    "pbs-installer (==2025.8.18)",
    "pdfminer-six (==20250506)",
    "pika (==1.3.2)",
    "pillow (==11.3.0)",
    "pip-autoremove (==0.10.0)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.8)",
    "poetry (==2.1.4)",
    "poetry-core (==2.1.3)",
    "prometheus-client (==0.22.1)",
    "prompt-toolkit (==3.0.51)",
    "psutil (==7.0.0)",
    "pure-eval (==0.2.3)",
    "pycparser (==2.22)",
    "pycryptodome (==3.23.0)",
    "pydantic (==2.11.7)",
    "pydantic-core (==2.33.2)",
    "pygments (==2.19.2)",
    "pyparsing (==3.2.3)",
    "pypdf2 (==3.0.1)",
    "pyproject-hooks (==1.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.1.1)",
    "python-json-logger (==3.3.0)",
    "python-magic (==0.4.27)",
    "pywavelets (==1.8.0)",
    "pywin32 (==311) ; sys_platform == 'win32'",
    "pywin32-ctypes (==0.2.3) ; sys_platform == 'win32'",
    "pywinpty (>=2.0.0) ; sys_platform == 'win32'",
    "pyyaml (==6.0.2)",
    "pyzmq (==27.0.1)",
    "rapidfuzz (==3.13.0)",
    "referencing (==0.36.2)",
    "regex (==2025.7.34)",
    "requests (==2.32.4)",
    "requests-toolbelt (==1.0.0)",
    "rfc3339-validator (==0.1.4)",
    "rfc3986-validator (==0.1.1)",
    "rfc3987-syntax (==1.1.0)",
    "rpds-py (==0.26.0)",
    "safetensors (==0.5.3)",
    "scikit-image (==0.19.3)",
    "scipy (==1.15.3)",
    "send2trash (==1.8.3)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "soupsieve (==2.7)",
    "stack-data (==0.6.3)",
    "sympy (==1.13.3)",
    "terminado (==0.18.1)",
    "tifffile (==2025.5.10)",
    "timm (==1.0.19)",
    "tinycss2 (==1.4.0)",
    "tokenizers (==0.21.4)",
    "tomli (==2.2.1)",
    "tomlkit (==0.13.3)",
    "tornado (==6.5.1)",
    "tqdm (==4.67.1)",
    "traitlets (==5.14.3)",
    "transformers (==4.54.1)",
    "trove-classifiers (==2025.8.6.13)",
    "types-python-dateutil (==2.9.0.20250708)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.14.1)",
    "uri-template (==1.3.0)",
    "urllib3 (==2.5.0)",
    "virtualenv (==20.32.0)",
    "wcwidth (==0.2.13)",
    "webcolors (==24.11.1)",
    "webencodings (==0.5.1)",
    "websocket-client (==1.8.0)",
    "zipp (==3.23.0)",
    "zstandard (==0.24.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
